chore: move addGovernorsToEconCharter to econ-behaviors.js
