+++ packages/inter-protocol/src/proposals/econ-behaviors.js
@@ -791,3 +791,44 @@ export const startStakeFactory = async (
   ]);
 };
 harden(startStakeFactory);
+
+/**
+ * Introduce charter to governed creator facets.
+ *
+ * XXX make this the responsibility of startVaultFactory, startReserve, etc.?
+ *
+ * @param {import('./econ-behaviors').EconomyBootstrapPowers} powers
+ */
+export const addGovernorsToEconCharter = async ({
+  consume: { reserveKit, vaultFactoryKit, econCharterKit, auctioneerKit },
+  instance: {
+    consume: { reserve, VaultFactory, auctioneer },
+  },
+}) => {
+  const { creatorFacet } = E.get(econCharterKit);
+
+  await Promise.all(
+    [
+      {
+        label: 'reserve',
+        instanceP: reserve,
+        facetP: E.get(reserveKit).governorCreatorFacet,
+      },
+      {
+        label: 'VaultFactory',
+        instanceP: VaultFactory,
+        facetP: E.get(vaultFactoryKit).governorCreatorFacet,
+      },
+      {
+        label: 'auctioneer',
+        instanceP: auctioneer,
+        facetP: E.get(auctioneerKit).governorCreatorFacet,
+      },
+    ].map(async ({ label, instanceP, facetP }) => {
+      const [instance, govFacet] = await Promise.all([instanceP, facetP]);
+
+      return E(creatorFacet).addInstance(instance, govFacet, label);
+    }),
+  );
+};
+harden(addGovernorsToEconCharter);
